refactor: Participant와 서브 타입 클래스 패키지 분리
