+++ src/main/java/controller/BlackJackController.java
@@ -1,10 +1,10 @@
 package controller;
 
-import domain.Dealer;
 import domain.Deck;
 import domain.GameResult;
-import domain.Player;
 import domain.cardsGenerator.RandomCardsGenerator;
+import domain.participant.Dealer;
+import domain.participant.Player;
 import java.util.List;
 import java.util.Map;
 import view.InputView;
