+++ src/main/java/domain/participant/Dealer.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.participant;
 
+import domain.Cards;
+import domain.GameResult;
 import java.util.Collections;
 import java.util.EnumMap;
 import java.util.HashMap;
